requirements.txt
pandas>=2.0.0
numpy>=1.24.3
scikit-learn>=1.0.2
streamlit>=1.24.0
matplotlib>=3.7.1
seaborn>=0.12.2
PyYAML>=6.0
-e .